Fixed download obstracles for airfield. Replace container to struct for parse airfield. Change type maps.

+++ mapview.h
@@ -7,8 +7,8 @@
 
 struct ObstraclePoint {
         QString id;
-        double lat;
-        double lon;
+        float lat;
+        float lon;
         QString height;
         bool marker;
         bool type;
